+++ packages/kit/test/apps/basics/test/server.test.js
@@ -304,7 +304,7 @@ test.describe('Routing', () => {
 	test('event.params are available in handle', async ({ request }) => {
 		const response = await request.get('/routing/params-in-handle/banana');
 		expect(await response.json()).toStrictEqual({
-			key: 'routing/params-in-handle/[x]',
+			key: '/routing/params-in-handle/[x]',
 			params: { x: 'banana' }
 		});
 	});